Expose some implementation of `builtins.derivationStrict` for unit testing
This allows us to avoid needing to generalize `dummy://` into a
`transient://`  at this time.

Setting the global settings in unit tests like this is very ugly, but
the real problem is not changing the settings but having the code being
unit tested depend on global variables at all. #5638 tracks fixing this.